Redirect to non-www via JS
This is bad for SEO, but since SEO is not of our concern(*), we can just do the redirect in JS. This follows the pattern for the non-http-redirect right above.

(*) With Google Webmaster Tools it would also be possible to specify a primary domain to solve the duplicate content problem.