Mobile: Fix vertical scrollbar on About page
The broken image with `width=600` caused a vertical scrollbar on mobile. Since I don't know a current image URL, I commented it out for now.

I could not find a way to test this locally, so this is only tested in the Browser inspector.

The `wmax720` apparently does nothing, so I removed it.