Mobile: Provide access to most feature
This does  a lot of stuff:

It renders the `NavbarChangeset` on Route /changesets for mobile just the way it is for desktop. This shows the Component on mobile.

Adjust the height of the Logo-Navbar and Changeset-Navbar (`h40`) on mobile. Make sure the Map still calculates to 100% of the rest of the screen.

Change the OpenIn Component, to link to OMS.org on mobile. On Desktop, this is a separate link but on mobile we don't have the space. Also, move the OpenIn-Dropdown-Button-Text to the calling view.

Tags and Verify Component use different CSS now base on flex to handle the edge case of long usernames which would break the `NavbarChangeset`.

Removed unused tag-indentation in the Verify Component. It's easier to style, if there are just the level of tags present that are required for the style. Apart from that, just some spacing adjustments.

Dropdown Component does not have a default `mr3`, but the calling components provide the margin if needed (mostly not).

Changeset-Row Component: Get rid of non-needed tag-hierarchy (same as above).

Navbar Component: Introduce classes for the title and buttons section to allow custom styles on the calling end. Adjust sizes for desktop/mobile.

Map Component: Remove non needed empty div hierarchy (see above).

Navbar Changeset Component: Remove currently not used old mobile-Link. Change styling for mobile quite a bit to get it all show in one line. Longer usernames are cut of with '…' (via CSS). The flex styling is a bit complex for those Navbars.
Use a different style for the `Changeset: <ID>` on mobile.

Navbar Sidebar Component: Remove unused local state. Introduce a Burger Menu Button next to the menu as an affordance on mobile that the Logo is actually the place to see something and not just a home-link.