Improve minimum char limit in location search
Make the minimum limit depend on the input characters.
No limit is used if the input contains characters that are used
in the following CJK scripts: Han, Hangul, Hiragana and Katakana.
These scripts were chosen because there are place names that can
be written with 1 character in Han, Hiragana and Katakana. In Hangul,
it looks possible that such place names exist and often an input
method similar to the other scripts is used, which reduces
the probability of triggering search with only 1 character
when the user has not yet finished typing. New scripts can be added
to the regular expression with a similar reasoning.

`scx` is used in the regular expression because `Script_Extensions`
is a more useful property than `Script` when a single Unicode character
can be used in multiple scripts.

It would be even simpler to just remove the limit completely because
Nominatim does not return prefix matches. However, at least queries
that contain only 1 Latin letter take longer than queries with at least
a few characters. Also, the component that displays the results seems
to have some kind of problem displaying the results for the subsequent
queries and this problem might become more apparent for more users
if the character limit is removed.