Fix bug preventing paginator from working on first load
The paginator widget in the changeset list sidebar had an odd behavior
where the prev/next arrows did not work after initial page load, and
would only start working once you clicked on a specific page number
(e.g '2'). Turns out this is because of some sneaky square brackets in
one of the Redux selectors which caused pageIndex to be accidentally
converted into an array `[0]`. The expression `[0] + 1` evaluates to
the string `"01"` in JavaScript (of course it does), resulting in the
frontend trying to fetch page `01`, then `011`, then `0111`, etc on each
subsequent click of the next button, none of which would succeed.